Refactor Grid class to use float for cellSize
Modified the Grid constructor to accept cellSize as a float instead of an int, updated the cellSize method to return a float, and changed the member variable m_cellSize from int to float. Changes applied to both Grid.cpp and Grid.h.

+++ Grid.cpp
@@ -3,7 +3,7 @@
 #include <unordered_map>
 #include <algorithm>
 
-Grid::Grid(int cellSize, int rows, int cols, float posX, float posY) :
+Grid::Grid(float cellSize, int rows, int cols, float posX, float posY) :
 	m_cellSize(cellSize),
 	m_rows(rows),
 	m_cols(cols),
